Moves player methods from GameMap to Player
