+++ src/controllers/main_controller.py
@@ -2,6 +2,7 @@ from .game_state import GameState
 from .intro_controller import IntroController
 from .travel_controller import TravelController
 from .battle_controller import BattleController
+from models.game_map import GameMap
 from utils.music import MusicPlayer, SoundPlayer
 
 
@@ -11,7 +12,8 @@ class AdventureGame:
   def __init__(self):
     self.game_state   = GameState.INTRO
     self.playing_game = True
-    self.game_map     = None  # Waits until player is created
+    self.game_map     = GameMap()
+    self.player       = None  # Waits until player is created
 
     self.intro        = IntroController(self)
     self.travel       = TravelController(self)
@@ -19,6 +21,7 @@ class AdventureGame:
 
     self.music_player = MusicPlayer()
     self.sound_player = SoundPlayer()
+    self.music_player.turn_off()
 
   def main(self):
     while self.playing_game:
@@ -31,5 +34,8 @@ class AdventureGame:
       elif self.game_state == GameState.BATTLE:
         self.battle.main()
 
+    # Once we're done playing,
+    # stop the music/sound players
+    # to exit peacefully
     self.music_player.stop_track()
     self.sound_player.stop_track()
