+++ src/models/room.py
@@ -8,24 +8,8 @@ class Room:
     self.description = description
     self.inv = inventory if inventory else []
 
-  def show_inv(self):
-    """
-    Outputs all item names from the inventory as a string, formatted like so:
 
-        [ (item1), (item2), (Big butter sword) ]
-
-    If inventory is empty, returns 
-        [ empty ]
-    """
-    out = ""
-    if len(self.inv) == 0:
-      out += "~xempty~e"
-    else:
-      item_names = [f"~c({str(item)})~e" for item in self.inv]
-      out += ", ".join(item_names)
-    return f"[ {out} ]"
-
-  def remove_item(self, name):
+  def remove_from_inv(self, name):
     """
     If possible, removes item from room.
     Casing does not matter.
@@ -41,13 +25,29 @@ class Room:
         return item
     return None
 
-  def add_item(self, item):
+  def add_to_inv(self, item):
     """Adds item to room."""
     self.inv.append(item)
 
   def __str__(self):
     return f"~W{self.name}\n~e~g{self.description}\n~eItems: {self.show_inv()}"
 
+  def deprecated_show_inv(self):
+    """
+    Outputs all item names from the inventory as a string, formatted like so:
+
+        [ (item1), (item2), (Big butter sword) ]
+
+    If inventory is empty, returns 
+        [ empty ]
+    """
+    out = ""
+    if len(self.inv) == 0:
+      out += "~xempty~e"
+    else:
+      item_names = [f"~c({str(item)})~e" for item in self.inv]
+      out += ", ".join(item_names)
+    return f"[ {out} ]"
 
 class DebugRoom(Room):
   def __init__(self):
